module-02-python-automation/requirements.txt
# Test Framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-rerunfailures==14.0

# API & HTTP
requests==2.31.0
httpx==0.25.2

# Data Validation
jsonschema==4.20.0
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
Faker==20.1.0
allure-pytest==2.13.2

# Reporting
pytest-html==4.1.1
pytest-json-report==1.5.0